+++ extension/src/content/handlers/pageWatchers/watchHomePage.ts
