+++ extension/src/content/handlers/pageWatchers/watchSubscriptionPage.ts
